+++ test/Fixture/Methods/NoParameterWithNullDefaultValueRule/Failure/MethodInClassWithParameterWithNullDefaultValue.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Localheinz\PHPStan\Rules\Test\Fixture\Methods\NoParameterWithNullDefaultValueRule\Failure;
+
+final class MethodInClassWithParameterWithNullDefaultValue
+{
+    public function foo($bar = null)
+    {
+        return $bar;
+    }
+}