+++ test/Fixture/Methods/NoParameterWithNullDefaultValueRule/Failure/MethodInInterfaceWithParameterWithWronlgyCapitalizedNullDefaultValue.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Localheinz\PHPStan\Rules\Test\Fixture\Methods\NoParameterWithNullDefaultValueRule\Failure;
+
+interface MethodInInterfaceWithParameterWithWronlgyCapitalizedNullDefaultValue
+{
+    public function foo($bar = NuLl);
+}