+++ test/Fixture/Methods/NoParameterWithNullDefaultValueRule/Success/MethodInTraitWithParameterWithRootNamespaceReferencedNullDefaultValue.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Localheinz\PHPStan\Rules\Test\Fixture\Methods\NoParameterWithNullDefaultValueRule\Success;
+
+trait MethodInTraitWithParameterWithRootNamespaceReferencedNullDefaultValue
+{
+    public function foo($bar = \null)
+    {
+        return $bar;
+    }
+}